requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
gunicorn==20.0.4